Added SharedSearchEntry struct Also changed the datatype of evaluation to i16 so it may more easily be encoded

+++ src/evaluate.rs
@@ -1,18 +1,18 @@
 use super::*;
 use position::Position;
 
-pub fn evaluate(pos: &Position) -> i32 {
+pub fn evaluate(pos: &Position) -> i16 {
     material(pos)
 }
 
 /// Calculate a game phase value to allow interpolation of middlegame and
 /// endgame phases. Middlegame 24 -> 0 Endgame
-fn _game_phase(pos: &Position) -> i32 {
-    const KNIGHT: i32 = 1;
-    const BISHOP: i32 = 1;
-    const ROOK: i32 = 2;
-    const QUEEN: i32 = 4;
-    const TOTAL: i32 = 24;
+fn _game_phase(pos: &Position) -> i16 {
+    const KNIGHT: i16 = 1;
+    const BISHOP: i16 = 1;
+    const ROOK: i16 = 2;
+    const QUEEN: i16 = 4;
+    const TOTAL: i16 = 24;
     // If phase is > 24, due to promotion, return phase at maximum value of 24
     std::cmp::min(
         KNIGHT * pos.data.knight_sum()
@@ -23,12 +23,12 @@ fn _game_phase(pos: &Position) -> i32 {
     )
 }
 
-fn material(pos: &Position) -> i32 {
-    const QUEEN: i32 = 1000;
-    const ROOK: i32 = 525;
-    const BISHOP: i32 = 350;
-    const KNIGHT: i32 = 350;
-    const PAWN: i32 = 100;
+fn material(pos: &Position) -> i16 {
+    const QUEEN: i16 = 1000;
+    const ROOK: i16 = 525;
+    const BISHOP: i16 = 350;
+    const KNIGHT: i16 = 350;
+    const PAWN: i16 = 100;
     QUEEN * pos.data.queen_diff()
         + ROOK * pos.data.rook_diff()
         + BISHOP * pos.data.bishop_diff()
